+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [3.4.0] - 2021-05-12
+
 ### Added
 - Support for Keycloak 13
   *Note*: If you get an error like `client already exists` or `java.lang.IllegalStateException: Session/EntityManager is closed`, it's not an error in keycloak-config-cli.
@@ -397,11 +399,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- @formatter:off -->
 
-[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc3...HEAD
 [3.4.0-rc3]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc2...v3.4.0-rc3
 [3.4.0-rc2]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc1...v3.4.0-rc2
 [3.4.0-rc1]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0-rc0...v3.4.0-rc1
 [3.4.0-rc0]: https://github.com/adorsys/keycloak-config-cli/compare/v3.3.1...v3.4.0-rc0
+[Unreleased]: https://github.com/adorsys/keycloak-config-cli/compare/v3.4.0...HEAD
+[3.4.0]: https://github.com/adorsys/keycloak-config-cli/compare/v3.3.1...v3.4.0
 [3.3.1]: https://github.com/adorsys/keycloak-config-cli/compare/v3.3.0...v3.3.1
 [3.3.0]: https://github.com/adorsys/keycloak-config-cli/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/adorsys/keycloak-config-cli/compare/v3.1.3...v3.2.0